+++ example/get_ios_dependencies.sh
@@ -7,12 +7,16 @@ wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/2.4.0/ios_depende
 unzip -u -o ios_dependencies.zip
 rm ios_dependencies.zip
 
-wget https://github.com/tuarua/AR-ANE/releases/download/0.1.0/ios_dependencies.zip
+wget https://github.com/tuarua/AR-ANE/releases/download/0.2.0/ios_dependencies.zip
 unzip -u -o ios_dependencies.zip
 rm ios_dependencies.zip
 
-wget https://github.com/tuarua/AR-ANE/releases/download/0.1.0/assets.zip
+wget https://github.com/tuarua/AR-ANE/releases/download/0.2.0/assets.zip
 unzip -u -o assets.zip
 rm assets.zip
 
-wget -O ../native_extension/ane/ARANE.ane https://github.com/tuarua/AR-ANE/releases/download/0.1.0/ARANE.ane?raw=true
+wget https://github.com/tuarua/AR-ANE/releases/download/0.2.0/reference_images.zip
+unzip -u -o reference_images.zip
+rm reference_images.zip
+
+wget -O ../native_extension/ane/ARANE.ane https://github.com/tuarua/AR-ANE/releases/download/0.2.0/ARANE.ane?raw=true