+++ native_library/ios/ARANE/ARANE_FW/Extensions/Geometry/FreSCNPyramid.swift
@@ -84,7 +84,7 @@ public extension SCNPyramid {
         }
     }
     
-    func toFREObject(nodeName: String?) -> FREObject? {
+    @objc override func toFREObject(nodeName: String?) -> FREObject? {
         do {
             let ret = try FREObject(className: "com.tuarua.arane.shapes.Pyramid")
             try ret?.setProp(name: "width", value: self.width.toFREObject())