+++ native_library/ios/ARANE/ARANE_FW/Scene3DVC+ARSCNViewDelegate.swift
@@ -41,12 +41,13 @@ extension Scene3DVC: ARSCNViewDelegate, ARSessionDelegate {
     }
     
     func renderer(_ renderer: SCNSceneRenderer, didAdd node: SCNNode, for anchor: ARAnchor) {
-        if listeners.contains(AREvent.ON_PLANE_DETECTED), planeDetection,
+        if listeners.contains(AREvent.ON_PLANE_DETECTED), hasPlaneDetection,
             let planeAnchor = anchor as? ARPlaneAnchor {
+            planeAnchors[planeAnchor.identifier.uuidString] = planeAnchor
             node.name = UUID().uuidString
             var props = [String: Any]()
             props["anchor"] = [
-                "alignment": 0,
+                "alignment": planeAnchor.alignment.rawValue,
                 "id": planeAnchor.identifier.uuidString,
                 "center": ["x": planeAnchor.center.x, "y": planeAnchor.center.y, "z": planeAnchor.center.z],
                 "extent": ["x": planeAnchor.extent.x, "y": planeAnchor.extent.y, "z": planeAnchor.extent.z],
@@ -55,13 +56,32 @@ extension Scene3DVC: ARSCNViewDelegate, ARSessionDelegate {
             props["node"] = ["id": node.name]
             sendEvent(name: AREvent.ON_PLANE_DETECTED, value: JSON(props).description)
         }
+        
+        if #available(iOS 11.3, *) {
+            if listeners.contains(AREvent.ON_IMAGE_DETECTED),
+                let imageAnchor = anchor as? ARImageAnchor {
+                let referenceImage = imageAnchor.referenceImage
+                var props = [String: Any]()
+                props["anchor"] = [
+                    "id": imageAnchor.identifier.uuidString,
+                    "name": referenceImage.name ?? "",
+                    "width": referenceImage.physicalSize.width,
+                    "height": referenceImage.physicalSize.height,
+                    "transform": imageAnchor.transformAsArray
+                ]
+                node.name = UUID().uuidString
+                props["node"] = ["id": node.name]
+                sendEvent(name: AREvent.ON_IMAGE_DETECTED, value: JSON(props).description)
+            }
+        }
+        
     }
     
     func renderer(_ renderer: SCNSceneRenderer, didUpdate node: SCNNode, for anchor: ARAnchor) {
         if listeners.contains(AREvent.ON_PLANE_UPDATED), let planeAnchor = anchor as? ARPlaneAnchor {
             var props = [String: Any]()
             props["anchor"] = [
-                "alignment": 0,
+                "alignment": planeAnchor.alignment.rawValue,
                 "id": planeAnchor.identifier.uuidString,
                 "center": ["x": planeAnchor.center.x, "y": planeAnchor.center.y, "z": planeAnchor.center.z],
                 "extent": ["x": planeAnchor.extent.x, "y": planeAnchor.extent.y, "z": planeAnchor.extent.z],
@@ -73,9 +93,10 @@ extension Scene3DVC: ARSCNViewDelegate, ARSessionDelegate {
     }
     
     func renderer(_ renderer: SCNSceneRenderer, didRemove node: SCNNode, for anchor: ARAnchor) {
-        if listeners.contains(AREvent.ON_PLANE_REMOVED), let _ = anchor as? ARPlaneAnchor {
+        if listeners.contains(AREvent.ON_PLANE_REMOVED), let planeAnchor = anchor as? ARPlaneAnchor {
             var props = [String: Any]()
             props["nodeName"] = node.name
+            planeAnchors.removeValue(forKey: planeAnchor.identifier.uuidString)
             sendEvent(name: AREvent.ON_PLANE_REMOVED, value: JSON(props).description)
         }
     }